NetStorage API: fixed compilation errors and warnings on Windows. JIRA: CXX-3663
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@57688 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/misc/netstorage/netstorage_impl.hpp
@@ -85,7 +85,7 @@ public:
         m_Dirty = true;
     }
 
-    size_t GetCacheChunkSize() const {return m_CacheChunkSize;}
+    Uint8 GetCacheChunkSize() const {return m_CacheChunkSize;}
 
     void SetTTL(Uint8 ttl)
     {
